azure_functions/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.23.0
azure-storage-blob
pandas
scikit-learn
matplotlib
seaborn
azure-identity==1.23.0